refactor: AsynchronousParticipantService 사용 안할 수 있도록 deprecated 추가
