+++ backend/src/main/java/com/woowacourse/moamoa/study/controller/StudyParticipantController.java
@@ -2,7 +2,7 @@ package com.woowacourse.moamoa.study.controller;
 
 import com.woowacourse.moamoa.auth.config.AuthenticatedMemberId;
 import com.woowacourse.moamoa.study.service.AsyncService;
-import com.woowacourse.moamoa.study.service.SynchronizedParticipantService;
+import com.woowacourse.moamoa.study.service.StudyParticipantService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -16,14 +16,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/studies/{study-id}/members")
 public class StudyParticipantController {
 
-    private final SynchronizedParticipantService synchronizedParticipantService;
+    private final StudyParticipantService studyParticipantService;
     private final AsyncService asyncService;
 
     @PostMapping
     public ResponseEntity<Void> participateStudy(
             @AuthenticatedMemberId final Long memberId, @PathVariable("study-id") final Long studyId
     ) {
-        synchronizedParticipantService.participateStudy(memberId, studyId);
+        studyParticipantService.participateStudy(memberId, studyId);
         asyncService.send(studyId);
         return ResponseEntity.noContent().build();
     }
@@ -32,7 +32,7 @@ public class StudyParticipantController {
     public ResponseEntity<Void> leaveStudy(
             @AuthenticatedMemberId final Long memberId, @PathVariable("study-id") final Long studyId
     ) {
-        synchronizedParticipantService.leaveStudy(memberId, studyId);
+        studyParticipantService.leaveStudy(memberId, studyId);
         return ResponseEntity.noContent().build();
     }
 
@@ -42,7 +42,7 @@ public class StudyParticipantController {
             @PathVariable("study-id") final Long studyId,
             @PathVariable("member-id") final Long participantId
     ) {
-        synchronizedParticipantService.kickOutMember(memberId, studyId, participantId);
+        studyParticipantService.kickOutMember(memberId, studyId, participantId);
         return ResponseEntity.noContent().build();
     }
 }
