+++ backend/src/main/java/com/woowacourse/moamoa/study/service/AsynchronousParticipantService.java
@@ -1,0 +1,51 @@
+package com.woowacourse.moamoa.study.service;
+
+import com.woowacourse.moamoa.common.utils.DateTimeSystem;
+import com.woowacourse.moamoa.member.domain.repository.MemberRepository;
+import com.woowacourse.moamoa.member.service.exception.MemberNotFoundException;
+import com.woowacourse.moamoa.study.domain.Participant;
+import com.woowacourse.moamoa.study.domain.Study;
+import com.woowacourse.moamoa.study.domain.repository.StudyRepository;
+import com.woowacourse.moamoa.study.service.exception.StudyNotFoundException;
+import java.time.LocalDate;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class AsynchronousParticipantService {
+
+    private final MemberRepository memberRepository;
+    private final StudyRepository studyRepository;
+    private final DateTimeSystem dateTimeSystem;
+
+    @Deprecated
+    public void participateStudy(final Long memberId, final Long studyId) {
+        memberRepository.findById(memberId)
+                .orElseThrow(MemberNotFoundException::new);
+        final Study study = studyRepository.findById(studyId)
+                .orElseThrow(StudyNotFoundException::new);
+
+        study.participate(memberId);
+    }
+
+    @Deprecated
+    public void leaveStudy(final Long memberId, final Long studyId) {
+        final Study study = studyRepository.findById(studyId)
+                .orElseThrow(StudyNotFoundException::new);
+
+        final LocalDate now = dateTimeSystem.now().toLocalDate();
+        study.leave(new Participant(memberId), now);
+    }
+
+    @Deprecated
+    public void kickOutMember(final Long ownerId, final Long studyId, final Long participantId) {
+        final Study study = studyRepository.findById(studyId)
+                .orElseThrow(StudyNotFoundException::new);
+
+        final LocalDate now = dateTimeSystem.now().toLocalDate();
+        study.kickOut(ownerId, new Participant(participantId), now);
+    }
+}
