+++ backend/src/main/java/com/woowacourse/moamoa/study/service/StudyParticipantService.java
@@ -1,48 +1,23 @@
 package com.woowacourse.moamoa.study.service;
 
-import com.woowacourse.moamoa.common.utils.DateTimeSystem;
-import com.woowacourse.moamoa.member.domain.repository.MemberRepository;
-import com.woowacourse.moamoa.member.service.exception.MemberNotFoundException;
-import com.woowacourse.moamoa.study.domain.Participant;
-import com.woowacourse.moamoa.study.domain.Study;
-import com.woowacourse.moamoa.study.domain.repository.StudyRepository;
-import com.woowacourse.moamoa.study.service.exception.StudyNotFoundException;
-import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@Transactional
 @RequiredArgsConstructor
 public class StudyParticipantService {
 
-    private final MemberRepository memberRepository;
-    private final StudyRepository studyRepository;
-    private final DateTimeSystem dateTimeSystem;
+    private final AsynchronousParticipantService asynchronousParticipantService;
 
-    public void participateStudy(final Long memberId, final Long studyId) {
-        memberRepository.findById(memberId)
-                .orElseThrow(MemberNotFoundException::new);
-        final Study study = studyRepository.findById(studyId)
-                .orElseThrow(StudyNotFoundException::new);
-
-        study.participate(memberId);
+    public synchronized void participateStudy(final Long memberId, final Long studyId) {
+        asynchronousParticipantService.participateStudy(memberId, studyId);
     }
 
-    public void leaveStudy(final Long memberId, final Long studyId) {
-        final Study study = studyRepository.findById(studyId)
-                .orElseThrow(StudyNotFoundException::new);
-
-        final LocalDate now = dateTimeSystem.now().toLocalDate();
-        study.leave(new Participant(memberId), now);
+    public synchronized void leaveStudy(final Long memberId, final Long studyId) {
+        asynchronousParticipantService.leaveStudy(memberId, studyId);
     }
 
-    public void kickOutMember(final Long ownerId, final Long studyId, final Long participantId) {
-        final Study study = studyRepository.findById(studyId)
-                .orElseThrow(StudyNotFoundException::new);
-
-        final LocalDate now = dateTimeSystem.now().toLocalDate();
-        study.kickOut(ownerId, new Participant(participantId), now);
+    public synchronized void kickOutMember(final Long ownerId, final Long studyId, final Long participantId) {
+        asynchronousParticipantService.kickOutMember(ownerId, studyId, participantId);
     }
 }
